refactor: 파싱 기능 제거
- Jackson 라이브러리가 자동적으로 타입을 변환해주기에 파싱 기능 및 테스트 삭제
