+++ src/main/java/roomescape/domain/reservations/entity/Reservation.java
@@ -10,19 +10,11 @@ public class Reservation {
     private LocalDate date;
     private LocalTime time;
 
-    public Reservation(Long id, String name, String date, String time) {
+    public Reservation(Long id, String name, LocalDate date, LocalTime time) {
         this.id = id;
         this.name = name;
-        this.date = parseLocalDate(date);
-        this.time = parseTime(time);
-    }
-
-    public LocalDate parseLocalDate(String date) {
-        return LocalDate.parse(date);
-    }
-
-    public LocalTime parseTime(String time) {
-        return LocalTime.parse(time);
+        this.date = date;
+        this.time = time;
     }
 
     public Long getId() {
